refactor: simplify to make cluster cert pkg codes more readable
