+++ pkg/clustercert/kubeconfig.go
@@ -12,16 +12,17 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package cert
+package clustercert
 
 import (
 	"bytes"
 	"crypto"
 	"crypto/x509"
 	"fmt"
-	"io"
 	"os"
 	"path/filepath"
+
+	"github.com/sealerio/sealer/pkg/clustercert/cert"
 
 	"github.com/pkg/errors"
 	"github.com/sirupsen/logrus"
@@ -54,10 +55,10 @@ type kubeConfigSpec struct {
 // join --control-plane workflow, plus the admin kubeconfig file used by the administrator and kubeadm itself; the
 // kubelet.conf file must not be created because it will be created and signed by the kubelet TLS bootstrap process.
 // If any kubeconfig files already exists, it used only if evaluated equal; otherwise an error is returned.
-func CreateJoinControlPlaneKubeConfigFiles(outDir string, cfg Config, nodeName, controlPlaneEndpoint, clusterName string) error {
+func CreateJoinControlPlaneKubeConfigFiles(outDir string, caCertPath, caCertName, nodeName, controlPlaneEndpoint, clusterName string) error {
 	return createKubeConfigFiles(
 		outDir,
-		cfg,
+		caCertPath, caCertName,
 		nodeName,
 		controlPlaneEndpoint,
 		clusterName,
@@ -68,18 +69,11 @@ func CreateJoinControlPlaneKubeConfigFiles(outDir string, cfg Config, nodeName, 
 	)
 }
 
-// cmd/kubeadm/app/cmd/phases/init/kubeconfig.go
-// cmd/kubeadm/app/phases/kubeconfig/kubeconfig.go
-func CreateKubeConfigFile(kubeConfigFileName string, outDir string, cfg Config, nodeName, controlPlaneEndpoint, clusterName string) error {
-	logrus.Infof("creating kubeconfig file for %s", kubeConfigFileName)
-	return createKubeConfigFiles(outDir, cfg, kubeConfigFileName, nodeName, controlPlaneEndpoint, clusterName)
-}
-
 // createKubeConfigFiles creates all the requested kubeconfig files.
 // If kubeconfig files already exists, they are used only if evaluated equal; otherwise an error is returned.
-func createKubeConfigFiles(outDir string, cfg Config, nodeName, controlPlaneEndpoint, clusterName string, kubeConfigFileNames ...string) error {
+func createKubeConfigFiles(outDir string, certPath, certName, nodeName, controlPlaneEndpoint, clusterName string, kubeConfigFileNames ...string) error {
 	// gets the KubeConfigSpecs, actualized for the current InitConfiguration
-	specs, err := getKubeConfigSpecs(cfg, nodeName, controlPlaneEndpoint)
+	specs, err := getKubeConfigSpecs(certPath, certName, nodeName, controlPlaneEndpoint)
 	if err != nil {
 		return err
 	}
@@ -108,10 +102,10 @@ func createKubeConfigFiles(outDir string, cfg Config, nodeName, controlPlaneEndp
 
 // getKubeConfigSpecs returns all KubeConfigSpecs actualized to the context of the current InitConfiguration
 // NB. these methods holds the information about how kubeadm creates kubeconfig files.
-func getKubeConfigSpecs(cfg Config, nodeName, controlPlaneEndpoint string) (map[string]*kubeConfigSpec, error) {
-	caCert, caKey, err := LoadCaCertAndKeyFromDisk(cfg)
+func getKubeConfigSpecs(certPath, certName, nodeName, controlPlaneEndpoint string) (map[string]*kubeConfigSpec, error) {
+	caCert, caKey, err := cert.NewCertificateFileManger(certPath, certName).Read()
 	if err != nil {
-		return nil, errors.Wrap(err, "couldn't create a kubeconfig; the CA files couldn't be loaded")
+		return nil, errors.Wrap(err, "couldn't create a kubeconfig; the CA cert file couldn't be loaded")
 	}
 
 	if len(nodeName) == 0 {
@@ -171,20 +165,26 @@ func buildKubeConfigFromSpec(spec *kubeConfigSpec, clustername string) (*clientc
 			spec.APIServer,
 			clustername,
 			spec.ClientName,
-			EncodeCertPEM(spec.CACert),
+			cert.EncodeCertPEM(spec.CACert),
 			spec.TokenAuth.Token,
 		), nil
 	}
 
 	// otherwise, create a client certs
-	clientCertConfig := Config{
+	clientCertConfig := cert.CertificateDescriptor{
 		CommonName:   spec.ClientName,
 		Organization: spec.ClientCertAuth.Organizations,
 		Usages:       []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth},
 		Year:         100,
 	}
 
-	clientCert, clientKey, err := NewCaCertAndKeyFromRoot(clientCertConfig, spec.CACert, spec.ClientCertAuth.CAKey)
+	g, err := cert.NewCommonCertificateGenerator(clientCertConfig, spec.CACert, spec.ClientCertAuth.CAKey)
+	if err != nil {
+		return nil, err
+	}
+
+	clientCert, clientKey, err := g.Generate()
+
 	if err != nil {
 		return nil, errors.Wrapf(err, "failure while creating %s client certificate", spec.ClientName)
 	}
@@ -198,9 +198,9 @@ func buildKubeConfigFromSpec(spec *kubeConfigSpec, clustername string) (*clientc
 		spec.APIServer,
 		clustername,
 		spec.ClientName,
-		EncodeCertPEM(spec.CACert),
+		cert.EncodeCertPEM(spec.CACert),
 		encodedClientKey,
-		EncodeCertPEM(clientCert),
+		cert.EncodeCertPEM(clientCert),
 	), nil
 }
 
@@ -275,104 +275,6 @@ func createKubeConfigFileIfNotExists(outDir, filename string, config *clientcmda
 	// kubeadm thinks those files are equal and doesn't bother writing a new file
 	logrus.Infof("[kubeconfig] Using existing kubeconfig file: %q\n", kubeConfigFilePath)
 
-	return nil
-}
-
-// WriteKubeConfigWithClientCert writes a kubeconfig file - with a client certificate as authentication info  - to the given writer.
-func WriteKubeConfigWithClientCert(out io.Writer, cfg Config, clientName, controlPlaneEndpoint, clusterName string, organizations []string) error {
-	// creates the KubeConfigSpecs, actualized for the current InitConfiguration
-	caCert, caKey, err := LoadCaCertAndKeyFromDisk(cfg)
-	if err != nil {
-		return errors.Wrap(err, "couldn't create a kubeconfig; the CA files couldn't be loaded")
-	}
-
-	if len(controlPlaneEndpoint) == 0 {
-		return errors.New("controlPlaneEndpoint  can not be empty")
-	}
-
-	spec := &kubeConfigSpec{
-		ClientName: clientName,
-		APIServer:  controlPlaneEndpoint,
-		CACert:     caCert,
-		ClientCertAuth: &clientCertAuth{
-			CAKey:         caKey,
-			Organizations: organizations,
-		},
-	}
-
-	return writeKubeConfigFromSpec(out, spec, clusterName)
-}
-
-// WriteKubeConfigWithToken writes a kubeconfig file - with a token as client authentication info - to the given writer.
-func WriteKubeConfigWithToken(out io.Writer, cfg Config, clientName, controlPlaneEndpoint, clusterName, token string) error {
-	// creates the KubeConfigSpecs, actualized for the current InitConfiguration
-	caCert, _, err := LoadCaCertAndKeyFromDisk(cfg)
-	if err != nil {
-		return errors.Wrap(err, "couldn't create a kubeconfig; the CA files couldn't be loaded")
-	}
-
-	if len(controlPlaneEndpoint) == 0 {
-		return errors.New("controlPlaneEndpoint  can not be empty")
-	}
-
-	spec := &kubeConfigSpec{
-		ClientName: clientName,
-		APIServer:  controlPlaneEndpoint,
-		CACert:     caCert,
-		TokenAuth: &tokenAuth{
-			Token: token,
-		},
-	}
-
-	return writeKubeConfigFromSpec(out, spec, clusterName)
-}
-
-// writeKubeConfigFromSpec creates a kubeconfig object from a kubeConfigSpec and writes it to the given writer.
-func writeKubeConfigFromSpec(out io.Writer, spec *kubeConfigSpec, clustername string) error {
-	// builds the KubeConfig object
-	config, err := buildKubeConfigFromSpec(spec, clustername)
-	if err != nil {
-		return err
-	}
-
-	// writes the kubeconfig to disk if it not exists
-	configBytes, err := clientcmd.Write(*config)
-	if err != nil {
-		return errors.Wrap(err, "failure while serializing admin kubeconfig")
-	}
-
-	fmt.Fprintln(out, string(configBytes))
-	return nil
-}
-
-// ValidateKubeconfigsForExternalCA check if the kubeconfig file exist and has the expected CA and server URL using kubeadmapi.InitConfiguration.
-func ValidateKubeconfigsForExternalCA(outDir string, cfg Config, controlPlaneEndpoint string) error {
-	kubeConfigFileNames := []string{
-		"admin.conf",
-		"kubelet.conf",
-		"controller-manager.conf",
-		"scheduler.conf",
-	}
-
-	// Creates a kubeconfig file with the target CA and server URL
-	// to be used as an input for validating user provided kubeconfig files
-	caCert, _, err := LoadCaCertAndKeyFromDisk(cfg)
-	if err != nil {
-		return err
-	}
-
-	if len(controlPlaneEndpoint) == 0 {
-		return errors.New("controlPlaneEndpoint  can not be empty")
-	}
-
-	validationConfig := CreateBasic(controlPlaneEndpoint, "dummy", "dummy", EncodeCertPEM(caCert))
-
-	// validate user provided kubeconfig files
-	for _, kubeConfigFileName := range kubeConfigFileNames {
-		if err = validateKubeConfig(outDir, kubeConfigFileName, validationConfig); err != nil {
-			return errors.Wrapf(err, "the %s file does not exists or it is not valid", kubeConfigFileName)
-		}
-	}
 	return nil
 }
